requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-multipart==0.0.6
openai-whisper==20231117
numpy==1.24.3
networkx==3.2.1
nltk==3.8.1
yt-dlp==2023.12.30
requests==2.31.0
supabase==2.0.3
python-dotenv==1.0.0
instaloader==4.10.2
ffmpeg-python==0.2.0
